add ci pipeline
The ci pipeline will react on pull requests and test the pull request
by deploying the gardener and creating a shoot on multiple providers.

merge.py can merge landscape_config.yaml and landscape_base.yaml files.
It was originally designed to be used as a git merge driver but is
currently not used in this way. It takes one "new" yaml and one "old"
one and fills the new one with the values from the old one. Merge
annotations can be used to finetune the merging behavior. They can be
provided as comments in the new yaml or in an extra file. The annotations
are necessary if the structure of the yaml changed between the old and
the new version.